Samples: Bump samples to SDK 2.9
The sample CaptureHDRCompleteSettings.hdev is complemented with the new
Region of Interest feature and the new filters: Cluster Filter and the
Hole Filling filter.

The procedure read_zivid_json_params.hdvp is also modified to support
the new Region of Interest feature and the new filters: Cluster Filter
and the Hole Filling filter.